moved check for exitst in top level regions from ExitValidator to RegionValidator
